requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pillow==11.0.0
numpy==2.1.3
